refactor: extract information types to domain values
- Move LocationInfo, InformationConflict, ConflictType, ConflictResolution to src/domain/values/information.rs
- Add re-exports in entities.rs for backward compatibility
- Separate domain value objects from infrastructure concerns
