+++ src/domain/values/mod.rs
@@ -1,3 +1,4 @@
+pub mod information;
 pub mod position;
 pub mod resource;
 pub mod terrain;
